--- src/example/myapp/Main.kt
@@ -1,39 +1,0 @@
-package example.myapp
-
-fun buildAquarium() {
-    val aquarium1 = Aquarium()
-    aquarium1.printSize()
-    // default height and length
-    val aquarium2 = Aquarium(width = 25)
-    aquarium2.printSize()
-    // default width
-    val aquarium3 = Aquarium(height = 35, length = 110)
-    aquarium3.printSize()
-    // everything custom
-    val aquarium4 = Aquarium(width = 25, height = 35, length = 110)
-    aquarium4.printSize()
-    println()
-    val aquarium6 = Aquarium(length = 25, width = 25, height = 40)
-    aquarium6.printSize()
-    println()
-    val myAquarium = Aquarium(width = 25, length = 25, height = 40)
-    myAquarium.printSize()
-    val myTower = TowerTank(diameter = 25, height = 40)
-    myTower.printSize()
-    println()
-}
-
-fun makeFish() {
-    val shark = Shark()
-    val pleco = Plecostomus()
-    println("Shark: ${shark.color}")
-    shark.eat()
-    println("Plecostomus: ${pleco.color}")
-    pleco.eat()
-    println()
-}
-
-fun main() {
-    buildAquarium()
-    makeFish()
-}
